Pass the verify_ssl instance variable to all requests
Previously, the passing of this variable was sporadic which makes
very little sense and causes issues when someone needs to communicate
with a server that doesn't have a secure certificate. Now all requests
should pass this variable allowing all functions to work with insecure
servers if the user requires it.