Add a convenience method to Get All results
The `getall()` function is implemented is a staticmethod, which can be
called by sending the gitlab function you want to use. It auto-paginates
by incrementing through `page` until there are no more `results` or when
`results` is boolean False (0 or False, for example). Which is as
returned by the API's original function and behaviour.